requirements.txt
opencv-python==4.9.0.80
mpu6050-raspberrypi
RPi.GPIO
loguru
